refactor(router-api): add exception handlers on daily category todo router api

+++ app/routers/daily.py
@@ -7,6 +7,7 @@ from models.daily import Daily
 from sqlalchemy.orm import Session
 from schema.daily.daily_request import DailyCreateRequest, DailyUpdateRequest
 from schema.daily.daily_response import DailyResponse
+from exceptions.handler import NotFoundException, ConflictException
 
 router = APIRouter(tags=["Daily"])
 
@@ -18,18 +19,15 @@ router = APIRouter(tags=["Daily"])
     summary = "전체 Daily 조회"
 )
 def get_dailies_handler(user_id: int, year: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
 
-    #존재하는 달력인지 검사 (404 NOT FOUND)
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
     existing_calendar = session.execute(
             select(Calendar).where(
                 Calendar.user_id == user_id,
@@ -38,12 +36,9 @@ def get_dailies_handler(user_id: int, year: int, session : Session = Depends(get
     ).scalar_one_or_none()
 
     if not existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 달력입니다."
-        )
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
-    #존재하는 날짜인지 검사 (404 NOT FOUND)
+    #존재하는 Daily 인지 검사 (404 NOT FOUND)
     existing_dailies = session.execute(
         select(Daily).where(
             Daily.calendar_id == existing_calendar.calendar_id
@@ -60,18 +55,15 @@ def get_dailies_handler(user_id: int, year: int, session : Session = Depends(get
     summary = "단일 Daily 조회"
 )
 def get_daily_handler(user_id: int, year: int, month: int, day: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
     
-    #존재하는 달력인지 검사 (404 NOT FOUND)
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
     existing_calendar = session.execute(
         select(Calendar).where(
             Calendar.user_id == user_id,
@@ -80,12 +72,9 @@ def get_daily_handler(user_id: int, year: int, month: int, day: int, session : S
     ).scalar_one_or_none()
 
     if not existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 달력입니다."
-        )
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
-    #존재하는 날짜인지 검사 (404 NOT FOUND)
+    #존재하는 Daily 인지 검사 (404 NOT FOUND)
     existing_daily = session.execute(
         select(Daily).where(
             Daily.calendar_id == existing_calendar.calendar_id,
@@ -95,10 +84,7 @@ def get_daily_handler(user_id: int, year: int, month: int, day: int, session : S
     ).scalar_one_or_none()
 
     if not existing_daily:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 날짜입니다."
-        )
+        raise NotFoundException("존재하지 않는 Daily 입니다.")
 
     return existing_daily
 
@@ -110,18 +96,15 @@ def get_daily_handler(user_id: int, year: int, month: int, day: int, session : S
     summary = "단일 Daily 생성"
 )
 def create_daily_handler(body: DailyCreateRequest, user_id: int, year: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
     
-    #존재하는 달력인지 검사 (404 NOT FOUND)
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
     existing_calendar = session.execute(
         select(Calendar).where(
             Calendar.year == year,
@@ -130,12 +113,9 @@ def create_daily_handler(body: DailyCreateRequest, user_id: int, year: int, sess
     ).scalar_one_or_none()
 
     if not existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 달력입니다."
-        )
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
-    #동일한 날짜가 존재하는지 중복 검사 (409 CONFLICT)
+    #동일한 Daily 가 존재하는지 중복 검사 (409 CONFLICT)
     existing_daily = session.execute(
         select(Daily).where(
             Daily.calendar_id == existing_calendar.calendar_id,
@@ -145,11 +125,9 @@ def create_daily_handler(body: DailyCreateRequest, user_id: int, year: int, sess
     ).scalar_one_or_none()
 
     if existing_daily:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail="이미 존재하는 날짜입니다."
-        )
+        raise ConflictException("이미 존재하는 Daily 입니다.")
 
+    #Daily 추가
     daily = Daily(
         calendar_id = existing_calendar.calendar_id,
         month = body.month,
@@ -169,18 +147,15 @@ def create_daily_handler(body: DailyCreateRequest, user_id: int, year: int, sess
     summary = "단일 Daily 삭제"
 )
 def delete_daily_handler(user_id: int, year: int, month: int, day: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
 
-    #존재하는 달력인지 검사 (404 NOT FOUND)
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
     existing_calendar = session.execute(
         select(Calendar).where(
             Calendar.year == year,
@@ -189,12 +164,9 @@ def delete_daily_handler(user_id: int, year: int, month: int, day: int, session 
     ).scalar_one_or_none()
 
     if not existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 달력입니다."
-        )
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
-    #존재하는 날짜인지 검사 (404 NOT FOUND)
+    #존재하는 Daily 인지 검사 (404 NOT FOUND)
     existing_daily = session.execute(
         select(Daily).where(
             Daily.calendar_id == existing_calendar.calendar_id,
@@ -204,11 +176,9 @@ def delete_daily_handler(user_id: int, year: int, month: int, day: int, session 
     ).scalar_one_or_none()
 
     if not existing_daily:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 날짜입니다."
-        )
+        raise NotFoundException("존재하지 않는 Daily 입니다.")
 
+    #Daily 삭제
     session.delete(existing_daily)
     session.commit()
 
@@ -221,18 +191,15 @@ def delete_daily_handler(user_id: int, year: int, month: int, day: int, session 
     summary = "전체 Daily 삭제"
 )
 def delete_dailies_handler(user_id: int, year: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
 
-    #존재하는 달력인지 검사 (404 NOT FOUND)
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
     existing_calendar = session.execute(
         select(Calendar).where(
             Calendar.year == year,
@@ -241,18 +208,19 @@ def delete_dailies_handler(user_id: int, year: int, session : Session = Depends(
     ).scalar_one_or_none()
 
     if not existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 달력입니다."
-        )
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
-    #존재하는 날짜인지 검사 (404 NOT FOUND)
+    #존재하는 Daily 인지 검사 (404 NOT FOUND)
     existing_dailies = session.execute(
         select(Daily).where(
             Daily.calendar_id == existing_calendar.calendar_id
         )
     ).scalars().all()
 
+    if not existing_dailies:
+        raise NotFoundException("존재하지 않는 Daily 입니다.")
+
+    #확인된 Dailies 삭제
     for existing_daily in existing_dailies:
         session.delete(existing_daily)
 
@@ -268,18 +236,15 @@ def delete_dailies_handler(user_id: int, year: int, session : Session = Depends(
     summary = "단일 Daily 수정"
 )
 def update_daily_handler(body: DailyUpdateRequest, user_id: int, year: int, month: int, day: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
     
-    #존재하는 달력인지 검사 (404 NOT FOUND)
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
     existing_calendar = session.execute(
         select(Calendar).where(
             Calendar.year == year,
@@ -288,12 +253,9 @@ def update_daily_handler(body: DailyUpdateRequest, user_id: int, year: int, mont
     ).scalar_one_or_none()
 
     if not existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 달력입니다."
-        )
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
-    #존재하는 날짜인지 검사 (404 NOT FOUND)
+    #존재하는 Daily 인지 검사 (404 NOT FOUND)
     existing_daily = session.execute(
         select(Daily).where(
             Daily.calendar_id == existing_calendar.calendar_id,
@@ -303,12 +265,9 @@ def update_daily_handler(body: DailyUpdateRequest, user_id: int, year: int, mont
     ).scalar_one_or_none()
 
     if not existing_daily:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 날짜입니다."
-        )
+        raise NotFoundException("존재하지 않는 Daily 입니다.")
 
-    #바꿀 날짜가 이미 존재하는 날짜인지 검사 (409 CONFLICT)
+    #바꿀 Daily 가 이미 존재하는 Daily 인지 검사 (409 CONFLICT)
     duplicate_daily = session.execute(
         select(Daily).where(
             Daily.calendar_id == existing_calendar.calendar_id,
@@ -318,11 +277,9 @@ def update_daily_handler(body: DailyUpdateRequest, user_id: int, year: int, mont
     ).scalar_one_or_none()
 
     if duplicate_daily:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail="이미 존재하는 날짜입니다."
-        )
+        raise ConflictException("이미 존재하는 Daily 입니다.")
 
+    #Daily 수정
     existing_daily.month = body.month
     existing_daily.day = body.day
